Generalize distribution summary
Sometimes it's useful to track the distribution summary of quantities
best represented by doubles. This generalizes the interface for
`distribution_summary`, and maintains source compatibility.  Since users
will have to manually create a distribution summary of type !=
`int64_t`, and then register it, we expose the required `pollers()`
method in the registry.