Allow opt-out of notifying alert-server
Currently the library notifies alert-server that we don't support
on-instance alerts. This was somewhat appropriate when there would be
only one atlas-client on the machine. Either the java client or this
native client, but now we have more users (like a system-agent) that
is likely to be installed alongside a jvm app, using the java client.

This PR adds a new function that can be used by system agents to opt-out
of notifying the alert server of lack of on-instance alerts.