Generate tables as part of the build
Previously we were relying on a checked in copy of the generated include
files, which could confuse someone thinking that by modifying the
generator source files it would have any effect. Files are now generated
as part of the build, so any changes made to the generator files will be
reflected in the input to the build.