Add support for test coverage reporting
Fixes #37. Currently we're doing a matrix build compiling for both g++
and clang++ and doing code coverage on both (meaning both are debug
builds with no optimizations). A refinement would be to do
it on only one of them, and do a release build for the other to verify
that the tests pass with optimizations as well.