Do not sleep before sending first batch of metrics
This fixes the case where the atlas client is initialized and stopped in
less time than it should take us to send the first batch of metrics.
Oridinarily we want to sleep a few seconds so we start collecting
metrics near the beginning of the step, but if we sleep then we cannot
notice the fact the user has requested a shutdown. Switch to the
condition variable that was introduced for this purpose.