Deal with compressed payload in HTTP GETs
libcurl was not automatically decompressing the payload. Manually handle
the case where the payload is compressed.

Note that we are allocating a fixed size 10MB buffer as the
max payload we can handle. This should be sufficient for now, but in the
near future we should dynamically resize the buffer as we uncompress the
data coming from network. (Switching to manual calls to inflate)