Use a vector as our buffer implementation
The previous buffer code was meant only to be used with curl/gzip which
pass big blocks of memory and therefore keeping the number of
reallocations/copies small. It relied on this fact to reduce to a
minimum the memory footprint (capacity=size).

Since buffer is now a more publicly visible class, then it makes sense
to use a more general purpose data structure.