Simplify API making it closer to Spectator
Many improvements to the API, plus remove support for multiple
frequencies for subscriptions. This is a step in taking us closer to the
Spectator AtlasRegistry so we can join our efforts in improving the
performance for higher frequency subscriptions.

As part of this we also remove the hack for 'check_cluster' which
notified the alert service that on-instance alerts are not supported by
our client. We also remove the `atlas_registry` global, and functions at
the top level namespace. Users should instantiate `atlas::Client` and
call Start/Stop methods once, and `GetRegistry()` to get the main
registry.

Other minor improvements plus fixes:

* Add a double backed distribution summary

* Add dedicated evaluator class and publisher classes to simplify
subscription_manager

* Make :by, etc., only return values if they have a number.

* Make :all take a query only so it returns all measurements matching the
query

* Create a singleton value expr to deal with simple value expressions in a
list context

* Add http.req.complete timer to http requests