Use address sanitizer on travis builds
When using travis enable the address sanitizer for clang builds. This
means we run two builds:

* gcc with code coverage and debug flags

* clang with address sanitizer checks and release flags

This should help us catch a wide variety of bugs, at the cost of a minor slow
down when running tests. When compiling this library for actual usage
ASAN should not be enabled to prevent a slow down.