Consolidation registry (#78)
Use a registry with a 5 second step, and aggregate those measurements
into the main registry (60 second). This allows us to support multiple
steps (as long as they are 5s multiples).

While testing subscriptions with multiple of 5s steps, I noticed we were
generating too many results. This was due to emitting metrics for
`,:count,...,:by` when count was 0. Now we no longer emit those 0
counts.