Handle registering the same Id with a diff type (#90)
This prevents the case where users would use the same `Id` for different
types. Our code would blindingly cast to the result type resulting in a
crash at runtime.

To achieve this we introduce a new method `GetClass()` to the `Meter`
interface. We cannot rely on RTTI on nodejs. As part of this refactor we
also remove the custom `SmallTagMap` class that was used to handle tags,
and use the `ska::flat_hash_map` that provides similar performance.
Since nodejs cannot work with code that uses exceptions I removed the
`at` methods and some `try` related code.