Use step numbers for the consolidated registry
We need to honor the timestamps and use step numbers to properly
attribute updates to a given interval. This also means we can't just
clear the registry after taken the measurements since the step numbers
can have partial updates for the next step, so we need a way to properly
expire them.