Use the beginning of the step for updates
Make the consolidated registry update its values based on the beginning
of the step instead of the end. And send values at the beginning of each
minute.

A simple test where we update a counter once per minute at each possible
second is added to ensure the basic expectation is met.