Use the beginning of the step for updates (#94)
Make the consolidated registry update its values based on the beginning
of the updating frequency step instead of the end. Send values at the beginning
of each minute.

A simple test where we update a counter once per minute at each possible
second is added to ensure the basic expectation is met.